pyproject.toml
[project]
name = "prompt-optimization-example"
version = "0.1.0"
description = "Automated prompt optimization using GEPA, Pydantic AI, and Pydantic Evals"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11",
    "pydantic-ai>=0.1.0",
    "pydantic-evals>=0.1.0",
    "gepa[full]>=0.0.24",
    "logfire>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_optimization"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
